GPT API 연동, 채팅 기능 일부 구현

+++ src/main/java/com/TingTing/repository/ConditionsRepository.java
@@ -3,6 +3,6 @@ package com.TingTing.repository;
 import com.TingTing.entity.Conditions;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConditionsRepository extends JpaRepository<Conditions, Long> {
+public interface ConditionsRepository extends JpaRepository<Conditions, Integer> {
 }
